refactor: move screens to a separate package
