+++ JetCatalogCompose/app/src/main/java/com/google/jetcatalog/screens/WorkInProgressScreen.kt
@@ -1,4 +1,4 @@
-package com.google.jetcatalog
+package com.google.jetcatalog.screens
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
